moved controls and brushes from theme library to Controls library * SymbolsControl * LineSeperator * CaretControl * Etc.
