+++ src/Consolonia.Controls/Brushes/LineStyle.cs
@@ -2,6 +2,9 @@
 
 namespace Consolonia.Controls.Brushes
 {
+    /// <summary>
+    /// Represents the style of a line.
+    /// </summary>
     public enum LineStyle
     {
         SingleLine = 0,
