+++ src/Consolonia.Controls/Brushes/LineStyles.cs
@@ -3,6 +3,9 @@ using System.Linq;
 
 namespace Consolonia.Controls.Brushes
 {
+    /// <summary>
+    /// Represents a LineStyle for each side of a rectangle.
+    /// </summary>
     public class LineStyles
     {
         public LineStyles()
