+++ src/Consolonia.Controls/Brushes/MoveConsoleCaretToPositionBrush.cs
@@ -1,8 +1,11 @@
 using Avalonia;
 using Avalonia.Media;
 
-namespace Consolonia.Core.Drawing
+namespace Consolonia.Controls.Brushes
 {
+    /// <summary>
+    /// This brush will move the console caret to the specified position it is drawn into.
+    /// </summary>
     public class MoveConsoleCaretToPositionBrush : AvaloniaObject, IImmutableBrush
     {
         public static readonly StyledProperty<CaretStyle> CaretStyleProperty =
