+++ src/Consolonia.Controls/Consolonia.Controls.csproj
@@ -5,6 +5,8 @@
     <ItemGroup>
         <PackageReference Include="Avalonia" Version="$(AvaloniaVersion)" />
         <PackageReference Include="CommunityToolkit.Mvvm" Version="8.3.2" />
+        <PackageReference Include="Unicode.net" Version="2.0.0" />
+        <PackageReference Include="Wcwidth" Version="2.0.0" />
     </ItemGroup>
 
 </Project>
