+++ src/Consolonia.Core/Drawing/PixelBufferImplementation/SimpleSymbol.cs
@@ -2,6 +2,7 @@ using System;
 using System.Diagnostics;
 using System.Diagnostics.CodeAnalysis;
 using System.Text;
+using Consolonia.Controls;
 using Consolonia.Core.Helpers;
 using Newtonsoft.Json;
