+++ src/Consolonia.Core/Dummy/DummyConsole.cs
@@ -1,4 +1,5 @@
 using Avalonia.Media;
+using Consolonia.Controls;
 using Consolonia.Core.Drawing;
 using Consolonia.Core.Drawing.PixelBufferImplementation;
 using Consolonia.Core.Infrastructure;
