+++ src/Consolonia.Core/Infrastructure/AnsiConsoleOutput.cs
@@ -2,6 +2,7 @@ using System;
 using System.Text;
 using Avalonia;
 using Avalonia.Media;
+using Consolonia.Controls;
 using Consolonia.Core.Drawing;
 using Consolonia.Core.Drawing.PixelBufferImplementation;
 using Consolonia.Core.Helpers;
