+++ src/Consolonia.Core/Infrastructure/DefaultNetConsoleOutput.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Text;
 using Avalonia.Media;
+using Consolonia.Controls;
 using Consolonia.Core.Drawing;
 using Consolonia.Core.Drawing.PixelBufferImplementation;
 using Consolonia.Core.Drawing.PixelBufferImplementation.EgaConsoleColor;
