+++ src/Consolonia.Core/Infrastructure/IConsoleOutput.cs
@@ -1,4 +1,5 @@
 using Avalonia.Media;
+using Consolonia.Controls;
 using Consolonia.Core.Drawing;
 using Consolonia.Core.Drawing.PixelBufferImplementation;
 
@@ -10,14 +11,9 @@ namespace Consolonia.Core.Infrastructure
     /// <summary>
     ///     Defines properties and methods for console output
     /// </summary>
-    public interface IConsoleOutput
+    public interface IConsoleOutput : IConsoleCapabilities
     {
         PixelBufferSize Size { get; set; }
-
-        /// <summary>
-        ///     This is true if console supports composing multiple emojis together (like: 👨‍👩‍👧‍👦).
-        /// </summary>
-        bool SupportsComplexEmoji { get; }
 
         /// <summary>
         ///     Set the title of the console window
