+++ src/Consolonia.Themes/Templates/Controls/Helpers/TextBoxCaret.cs
@@ -3,6 +3,7 @@ using Avalonia.Controls;
 using Avalonia.Media;
 using Avalonia.Media.Immutable;
 using Avalonia.Reactive;
+using Consolonia.Controls.Brushes;
 using Consolonia.Core.Drawing;
 
 namespace Consolonia.Themes.Templates.Controls.Helpers
