+++ src/Consolonia.Themes/Templates/Controls/SymbolsControl.axaml
@@ -1,8 +1,8 @@
 <ResourceDictionary xmlns="https://github.com/avaloniaui"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:helpers1="clr-namespace:Consolonia.Themes.Templates.Controls.Helpers;assembly=Consolonia.Themes">
-    <ControlTheme x:Key="{x:Type helpers1:SymbolsControl}"
-                  TargetType="helpers1:SymbolsControl">
+                    xmlns:console="https://github.com/jinek/consolonia">
+    <ControlTheme x:Key="{x:Type console:SymbolsControl}"
+                  TargetType="console:SymbolsControl">
         <!--<Setter Property="Foreground"
                 Value="{DynamicResource ThemeForegroundBrush}" />-->
     </ControlTheme>
