+++ src/Tests/Consolonia.Core.Tests/WithLifetimeFixture/GlyphTests.cs
@@ -1,3 +1,4 @@
+using Consolonia.Controls;
 using Consolonia.Core.Helpers;
 using NUnit.Framework;
